separated account info from home component
