+++ examples/wallet/app/Routes.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { Switch, Route } from 'react-router';
 import routes from './constants/routes';
 import App from './containers/App';
-import Home from './containers/Home';
+import Home from './components/Home';
 import Index from './containers/Index';
 import InputKeys from './containers/InputKeys';
