+++ examples/wallet/app/components/AccountInfo.js
@@ -1,0 +1,16 @@
+// @flow
+import React from 'react';
+
+type Props = {
+  balance: number,
+  address: string
+};
+
+export default ({ balance, address }: Props) => {
+  return (
+    <div>
+      <p>Current Address: {address}</p>
+      <p>Balance: {balance}</p>
+    </div>
+  );
+};
